refactor: 임시 User 엔티티 제거 및 리뷰 서비스 로직 정비
- 임시 User 엔티티 제거 및 실제 엔티티 적용
- ReviewService 내 DefaultException 오타 수정 (DefaultExeption → DefaultException)
- ReviewService 내 61번 라인 리뷰 작성자 이름 반환 시 getName() → getUsername()으로 수정
